pyproject.toml
[project]
name = "homelio"
version = "0.1.0"
description = "Local toolkit for deterministic tender document autofill."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.0",
    "python-docx>=1.2.0",
    "PyYAML>=6.0.2",
    "typer>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.12.0",
]

[project.scripts]
homelio = "homelio.cli:app"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

